PC-567: Fix update to multiply udf, add divide udf int/int->float
Summary: tsia.

Test Plan: skaffold (unit test for divide already exists)

Reviewers: zasgar, philkuz, jamesbartlett, #engineering

Reviewed By: zasgar, philkuz, #engineering

JIRA Issues: PC-567

Differential Revision: https://phab.corp.pixielabs.ai/D6418

GitOrigin-RevId: 272fa7d754dab56b23dd9ad717e7fd0b11665e43

+++ src/carnot/funcs/builtins/math_ops.cc
@@ -53,7 +53,7 @@ void RegisterMathOpsOrDie(udf::Registry* registry) {
           "subtract");
 
   // Division
-  registry->RegisterOrDie<DivideUDF<types::Int64Value, types::Int64Value, types::Int64Value>>(
+  registry->RegisterOrDie<DivideUDF<types::Float64Value, types::Int64Value, types::Int64Value>>(
       "divide");
   registry->RegisterOrDie<DivideUDF<types::Float64Value, types::Float64Value, types::Int64Value>>(
       "divide");
@@ -63,7 +63,7 @@ void RegisterMathOpsOrDie(udf::Registry* registry) {
       "divide");
 
   // Multiplication
-  registry->RegisterOrDie<MultiplyUDF<types::Float64Value, types::Int64Value, types::Int64Value>>(
+  registry->RegisterOrDie<MultiplyUDF<types::Int64Value, types::Int64Value, types::Int64Value>>(
       "multiply");
   registry->RegisterOrDie<MultiplyUDF<types::Float64Value, types::Float64Value, types::Int64Value>>(
       "multiply");
